requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
PyJWT>=2.8.0